Update: move Mutation() to a more reasonable place
