+++ enzy_htp/__init__.py
@@ -40,6 +40,8 @@ from .structure import (
     Mol2Parser
 )
 
+from .mutation_class import Mutation
+
 from .electronic_structure import ElectronicStructure
 
 from ._interface import Interface, OpenPyMolSession
@@ -53,6 +55,8 @@ config.__doc__="""Singleton interface for the config settings for all software p
 by end users and should be used to customize system settings for each environment they are working in."""
 
 from .preparation import PDBLine, read_pdb_lines, dock_reactants, generate_conformers #TODO(CJ): get rid of system charge
+
+from .mutation import assign_mutant, mutate_stru
 
 from .geometry import equi_md_sampling, md_simulation
