+++ enzy_htp/analysis/electric_field.py
@@ -21,7 +21,7 @@ import numpy as np
 
 from enzy_htp.core import _LOGGER
 from enzy_htp.chemical import electric_field_strength
-from enzy_htp.mutation import Mutation
+from enzy_htp.mutation_class import Mutation
 from enzy_htp.structure import Structure, Atom
 
 from enzy_htp.structure.structure_operation.charge import init_charge
