+++ enzy_htp/mutation/__init__.py
@@ -3,17 +3,9 @@ mutation module for enzy_htp.
 
 
 Author: Chris Jurich <chris.jurich@vanderbilt.edu>
+Author: QZ Shao <shaoqz@icloud.com>
 Date: 2022-06-15
 """
-
-from .mutation import (
-    Mutation,
-    generate_from_mutation_flag,
-    check_repeat_mutation,
-    remove_repeat_mutation,
-    get_mutant_name_tag,
-    get_involved_mutation,
-)
 
 from .api import (
     assign_mutant,
