+++ enzy_htp/mutation/mutation_pattern/api.py
@@ -13,7 +13,11 @@ from enzy_htp.core.exception import InvalidMutationPatternSyntax
 from enzy_htp.core.logger import _LOGGER
 from enzy_htp.core.general import (get_random_list_elem, pop_random_list_elem, product_lists_allow_empty, split_but_brackets)
 from enzy_htp.structure import Structure
-from ..mutation import (Mutation, generate_from_mutation_flag, generate_mutation_from_traget_list)
+from enzy_htp.mutation_class import (
+    Mutation, 
+    generate_from_mutation_flag, 
+    generate_mutation_from_traget_list
+)
 from .position_pattern import decode_position_pattern
 from .target_aa_pattern import check_target_aa_pattern, decode_target_aa_pattern
