+++ enzy_htp/mutation_class/__init__.py
@@ -1,0 +1,16 @@
+"""
+Mutation class for enzy_htp.
+
+
+Author: QZ Shao <shaoqz@icloud.com>
+Date: 2024-02-13
+"""
+
+from .mutation import (
+    Mutation,
+    generate_from_mutation_flag,
+    check_repeat_mutation,
+    remove_repeat_mutation,
+    get_mutant_name_tag,
+    get_involved_mutation,
+)
